sql拦截器还没有生效，准备工作已完成

+++ mybatisinterceptor/src/main/java/com/zy/mybatisinterceptor/dao/model/UserDo.java
@@ -18,10 +18,10 @@ import java.util.Date;
 @NoArgsConstructor
 public class UserDo implements Serializable {
 
-    private Byte id;
+    private int id;
     private String userName;
     private String sex;
-    private Byte age;
+    private int age;
     private String phoneNumber;
     private String address;
     private String pwd;
